Separate SourceIcon component
